+++ jest.config.server.js
@@ -2,9 +2,7 @@ const sharedConfig = require('./scripts/jest/jest.config.base.js');
 
 const moduleNameMapper = {
   '^(\\.{1,2}/.*)\\.js$': '$1',
-  '^lodash-es$': 'lodash',
-  '@jpmorganchase/mosaic-source-storybook/dist/transformer.js':
-    '<rootDir>/packages/source-storybook/dist/transformer.js'
+  '^lodash-es$': 'lodash'
 };
 
 module.exports = {