+++ packages/site-components/src/TableOfContents/TableOfContents.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useLayoutEffect, useRef, useState } from 'react';
+import React, { useEffect, useLayoutEffect, useRef, useState, useId } from 'react';
 import { throttle } from 'lodash-es';
 import { Caption1, useSize } from '@jpmorganchase/mosaic-components';
 
@@ -93,16 +93,16 @@ export const TableOfContents: React.FC<TableOfContentsProps> = ({ items }) => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [items]);
 
+  const captionId = useId();
+
   return items.length ? (
-    <nav>
-      <Caption1>On this page</Caption1>
-      <ul aria-label="Table of contents" className={styles.list} role="tree">
-        {items.map((item, i) => (
-          <TableOfContentsItem
-            current={selectedHeading}
-            item={item}
-            key={`TableOfContentsItem_${i}`}
-          />
+    <nav data-testid="table-of-contents" aria-labelledby={captionId}>
+      <Caption1 id={captionId}>On this page</Caption1>
+      <ul className={styles.list}>
+        {items.map(item => (
+          <li key={item.id}>
+            <TableOfContentsItem current={selectedHeading} item={item} />
+          </li>
         ))}
       </ul>
     </nav>