+++ packages/source-figma/src/__tests__/index.test.ts
@@ -72,7 +72,6 @@ const createProjectFilesResponse = (patternId: string, nodeId: string) => ({
 const createExpectedResult = (patternId: string, data: Record<string, any>) => ({
   title: patternId,
   description: `some description for ${patternId}`,
-  layout: 'DetailTechnical',
   route: `/prefixdir/${patternId}`.toLowerCase(),
   fullPath: `/prefixdir/${patternId}.json`.toLowerCase(),
   tags: ['some-tag1', 'some-tag2'],
@@ -84,8 +83,6 @@ const createExpectedResult = (patternId: string, data: Record<string, any>) => (
     nodeId: '2:0',
     name: patternId,
     patternId: patternId,
-    source: 'FIGMA',
-    tags: 'some-tag1,some-tag2',
     version: 'some version',
     ...data.data
   },
@@ -140,7 +137,7 @@ describe('GIVEN a Figma Source ', () => {
           };
           meta0.data = {
             ...meta0.data,
-            thumbnailUrl: `/thumbnail/file888/2:0`,
+            contentUrl: `/thumbnail/file888/2:0`,
             fileId: 'file888',
             projectId: '888'
           };
@@ -152,7 +149,7 @@ describe('GIVEN a Figma Source ', () => {
           meta1.data = {
             ...meta1.data,
             fileId: 'file999',
-            thumbnailUrl: `/thumbnail/file999/2:0`,
+            contentUrl: `/thumbnail/file999/2:0`,
             projectId: '999'
           };
           expect(result[1]).toEqual(createExpectedResult('jpmSaltPattern_999_pattern2', meta1));