+++ packages/source-storybook/src/__tests__/index.test.ts
@@ -3,22 +3,22 @@ import { setupServer } from 'msw/node';
 import { rest } from 'msw';
 import { StorybookPage } from '../types/index.js';
 
-import Source from '../index.js';
+import Source, { StorybookSourceOptions } from '../index.js';
 
 const schedule = {
   checkIntervalMins: 3,
   initialDelayMs: 0
 };
 
-const options = {
+const options: StorybookSourceOptions = {
   prefixDir: 'prefixDir',
   requestTimeout: 1000,
   stories: [
     {
       description: 'some description 1',
-      url: 'https://api.endpoint.com/1',
-      additionalData: { owner: 'some owner 1' },
-      additionalTags: ['some-additional-tag-1'],
+      storiesUrl: 'https://api.endpoint.com/1',
+      storyUrlPrefix: 'https://storybook.endpoint.com/1',
+      meta: { tags: ['some-additional-tag-1'], data: { owner: 'some owner 1' } },
       filter: /TestComponent\/SomePath\/Component/
     },
     /**
@@ -26,9 +26,9 @@ const options = {
      */
     {
       description: 'some description 2',
-      url: 'https://api.endpoint.com/2',
-      additionalData: { owner: 'some owner 2' },
-      additionalTags: ['some-additional-tag-2'],
+      storiesUrl: 'https://api.endpoint.com/2',
+      storyUrlPrefix: 'https://storybook.endpoint.com/2',
+      meta: { tags: ['some-additional-tag-2'], data: { owner: 'some owner 2' } },
       filterTags: ['tag-2']
     },
     /**
@@ -36,16 +36,16 @@ const options = {
      */
     {
       description: 'some description 3',
-      url: 'https://api.endpoint.com/3',
-      additionalData: { owner: 'some owner 3' },
-      additionalTags: ['some-additional-tag-3'],
+      storiesUrl: 'https://api.endpoint.com/3',
+      storyUrlPrefix: 'https://storybook.endpoint.com/3',
+      meta: { tags: ['some-additional-tag-3'], data: { owner: 'some owner 3' } },
       filter: /IgnoredComponent\/SomePath\/Component/
     },
     {
       description: 'some description 4',
-      url: 'https://api.endpoint.com/4',
-      additionalData: { owner: 'some owner 4' },
-      additionalTags: ['some-additional-tag-4'],
+      storiesUrl: 'https://api.endpoint.com/4',
+      storyUrlPrefix: 'https://storybook.endpoint.com/4',
+      meta: { tags: ['some-additional-tag-4'], data: { owner: 'some owner 4' } },
       filter: /TestComponent\/SomePath\/Component/
     },
     /**
@@ -53,7 +53,8 @@ const options = {
      */
     {
       description: 'some description 5',
-      url: 'https://api.endpoint.com/5',
+      storiesUrl: 'https://api.endpoint.com/5',
+      storyUrlPrefix: 'https://storybook.endpoint.com/5',
       filter: /TestComponent\/SomePath\/Component/
     }
   ]
@@ -76,37 +77,35 @@ const createResponse = (index: number) => ({
 
 const createExpectedResult = (index: number) => ({
   title: `TestComponent/SomePath/Component-${index} - Component ${index} Name`,
-  layout: 'DetailTechnical',
   route: `prefixDir/component${index}Id`,
-  fullPath: `prefixDir/component${index}Id.mdx`,
+  fullPath: `prefixDir/component${index}Id.json`,
   tags: [`some-additional-tag-${index}`],
   data: {
     id: `component${index}Id`,
+    contentUrl: `https://storybook.endpoint.com/${index}/iframe.html?id=component${index}Id&viewMode=story&shortcuts=false&singleStory=true`,
     description: `some description ${index}`,
     kind: `TestComponent/SomePath/Component-${index}`,
-    link: `https://api.endpoint.com/${index}/iframe.html?id=component${index}Id&viewMode=story&shortcuts=false&singleStory=true`,
+    link: `https://storybook.endpoint.com/${index}?id=component${index}Id`,
     name: `Component ${index} Name`,
     owner: `some owner ${index}`,
-    source: 'STORYBOOK',
     story: `Docs ${index}`
-  },
-  content: ''
+  }
 });
 
 const successHandlers = [
-  rest.get(`${options.stories[0].url}/stories.json`, (_req, res, ctx) => {
+  rest.get(options.stories[0].storiesUrl, (_req, res, ctx) => {
     return res(ctx.status(200), ctx.json(createResponse(1)));
   }),
-  rest.get(`${options.stories[1].url}/stories.json`, (_req, res, ctx) => {
+  rest.get(options.stories[1].storiesUrl, (_req, res, ctx) => {
     return res(ctx.status(200), ctx.json(createResponse(2)));
   }),
-  rest.get(`${options.stories[2].url}/stories.json`, (_req, res, ctx) => {
+  rest.get(options.stories[2].storiesUrl, (_req, res, ctx) => {
     return res(ctx.status(200), ctx.json(createResponse(3)));
   }),
-  rest.get(`${options.stories[3].url}/stories.json`, (_req, res, ctx) => {
+  rest.get(options.stories[3].storiesUrl, (_req, res, ctx) => {
     return res(ctx.status(200), ctx.json(createResponse(4)));
   }),
-  rest.get(`${options.stories[4].url}/stories.json`, (_req, res, ctx) => {
+  rest.get(options.stories[4].storiesUrl, (_req, res, ctx) => {
     return res(ctx.status(200), ctx.json(createResponse(5)));
   })
 ];
@@ -141,8 +140,8 @@ describe('GIVEN a Storybook Source ', () => {
           expect(result[1]).toEqual(createExpectedResult(4));
 
           const result5 = createExpectedResult(5);
-          result5.tags = []; // no tags added for this 1
-          delete result5.data.owner; // no additional data added for this 1
+          delete result5.tags; // no tags data added for this result5
+          delete result5.data.owner; // no additional data added for result5
           expect(result[2]).toEqual(result5);
         },
         complete: () => done()