+++ packages/source-storybook/src/index.ts
@@ -8,8 +8,8 @@ import {
   createProxyAgent
 } from '@jpmorganchase/mosaic-source-http';
 
-import createStorybookPages from './transformer.js';
-import { StoriesResponseJSON, StorybookPage } from './types/index.js';
+import { StoriesResponseJSON, StorybookPage, StoryConfig } from './types/index.js';
+import deepmerge from 'deepmerge';
 
 const baseSchema = httpSourceCreatorSchema.omit({
   endpoints: true, // will be generated from the url in the stories object,
@@ -23,10 +23,12 @@ export const schema = baseSchema.merge(
       .array(
         z.object({
           description: z.string(),
-          url: z.string(),
+          storiesUrl: z.string().url().optional(),
+          storyUrlPrefix: z.string().url(),
           proxyEndpoint: z.string().url().optional(),
-          additionalData: z.object({}).passthrough().optional(),
-          additionalTags: z.array(z.string()).optional(),
+          meta: z
+            .object({ tags: z.array(z.string()).optional(), data: z.object({}).passthrough() })
+            .optional(),
           filter: z
             .any()
             .transform(val => new RegExp(val))
@@ -40,6 +42,46 @@ export const schema = baseSchema.merge(
 
 export type StorybookSourceOptions = z.infer<typeof schema>;
 
+const transformStorybookPages = (
+  storyJSON: StoriesResponseJSON,
+  prefixDir: string,
+  index: number,
+  storyConfig: StoryConfig[]
+): StorybookPage[] => {
+  const { meta = {}, description, filter, filterTags, storyUrlPrefix } = storyConfig[index];
+  const storyIds = Object.keys(storyJSON.stories);
+  return storyIds.reduce<StorybookPage[]>((result, storyId) => {
+    const story = storyJSON.stories[storyId];
+    if (filter && !filter.test(story.kind)) {
+      return result;
+    }
+    if (filterTags && filterTags.some(filterTag => story.tags.indexOf(filterTag) >= 0)) {
+      return result;
+    }
+    const { id, kind, name, story: storyName } = story;
+    const title = `${kind} - ${name}`;
+    const route = `${prefixDir}/${id}`;
+    let storyPageMeta: StorybookPage = {
+      title,
+      route,
+      fullPath: `${route}.json`,
+      data: {
+        id,
+        description,
+        kind,
+        contentUrl: `${storyUrlPrefix}/iframe.html?id=${id}&viewMode=story&shortcuts=false&singleStory=true`,
+        link: `${storyUrlPrefix}?id=${id}`,
+        name,
+        story: storyName
+      }
+    };
+    if (meta) {
+      storyPageMeta = deepmerge<StorybookPage, Partial<StorybookPage>>(storyPageMeta, meta);
+    }
+    return [...result, storyPageMeta];
+  }, []);
+};
+
 const StorybookSource: Source<StorybookSourceOptions, StorybookPage> = {
   create(options, sourceConfig) {
     const parsedOptions = validateMosaicSchema(schema, options);
@@ -52,16 +94,16 @@ const StorybookSource: Source<StorybookSourceOptions, StorybookPage> = {
     } = parsedOptions;
 
     const configuredRequests = storiesConfig.map(config => {
-      const { url, proxyEndpoint } = config;
+      const { storiesUrl, storyUrlPrefix, proxyEndpoint } = config;
       let agent;
       const headers = requestHeaders ? (requestHeaders as HeadersInit) : undefined;
 
       if (proxyEndpoint) {
-        console.log(`[Mosaic] Storybook source using ${proxyEndpoint} proxy for ${url}`);
+        console.log(`[Mosaic] Storybook source using ${proxyEndpoint} proxy for ${storiesUrl}`);
         agent = createProxyAgent(proxyEndpoint);
       }
-
-      return new Request(`${url}/stories.json`, {
+      const url = storiesUrl || `${storyUrlPrefix}/stories.json`;
+      return new Request(url, {
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore
         agent,
@@ -75,7 +117,7 @@ const StorybookSource: Source<StorybookSourceOptions, StorybookPage> = {
         prefixDir,
         ...restOptions,
         configuredRequests,
-        transformer: createStorybookPages,
+        transformer: transformStorybookPages,
         transformerOptions: storiesConfig
       },
       sourceConfig