+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @jpmorganchase/mosaic-cli
 
+## 0.1.0-beta.69
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-core@0.1.0-beta.69
+- @jpmorganchase/mosaic-plugins@0.1.0-beta.69
+- @jpmorganchase/mosaic-source-local-folder@0.1.0-beta.69
+- @jpmorganchase/mosaic-standard-generator@0.1.0-beta.69
+
+## 0.1.0-beta.68
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-core@0.1.0-beta.68
+- @jpmorganchase/mosaic-plugins@0.1.0-beta.68
+- @jpmorganchase/mosaic-source-local-folder@0.1.0-beta.68
+- @jpmorganchase/mosaic-standard-generator@0.1.0-beta.68
+
+## 0.1.0-beta.67
+
+### Patch Changes
+
+- Updated dependencies [6abf1910]
+- Updated dependencies [30e2f038]
+  - @jpmorganchase/mosaic-plugins@0.1.0-beta.67
+  - @jpmorganchase/mosaic-core@0.1.0-beta.67
+  - @jpmorganchase/mosaic-source-local-folder@0.1.0-beta.67
+  - @jpmorganchase/mosaic-standard-generator@0.1.0-beta.67
+
+## 0.1.0-beta.66
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-core@0.1.0-beta.66
+- @jpmorganchase/mosaic-plugins@0.1.0-beta.66
+- @jpmorganchase/mosaic-source-local-folder@0.1.0-beta.66
+- @jpmorganchase/mosaic-standard-generator@0.1.0-beta.66
+
+## 0.1.0-beta.65
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-core@0.1.0-beta.65
+- @jpmorganchase/mosaic-plugins@0.1.0-beta.65
+- @jpmorganchase/mosaic-source-local-folder@0.1.0-beta.65
+- @jpmorganchase/mosaic-standard-generator@0.1.0-beta.65
+
 ## 0.1.0-beta.64
 
 ### Patch Changes