+++ packages/create-site/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @jpmorganchase/mosaic-create-site
 
+## 0.1.0-beta.69
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-standard-generator@0.1.0-beta.69
+
+## 0.1.0-beta.68
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-standard-generator@0.1.0-beta.68
+
+## 0.1.0-beta.67
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-standard-generator@0.1.0-beta.67
+
+## 0.1.0-beta.66
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-standard-generator@0.1.0-beta.66
+
+## 0.1.0-beta.65
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-standard-generator@0.1.0-beta.65
+
 ## 0.1.0-beta.64
 
 ### Patch Changes