+++ packages/source-http/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @jpmorganchase/mosaic-source-http
 
+## 0.1.0-beta.69
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-from-http-request@0.1.0-beta.69
+- @jpmorganchase/mosaic-schemas@0.1.0-beta.69
+- @jpmorganchase/mosaic-types@0.1.0-beta.69
+
+## 0.1.0-beta.68
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-from-http-request@0.1.0-beta.68
+- @jpmorganchase/mosaic-schemas@0.1.0-beta.68
+- @jpmorganchase/mosaic-types@0.1.0-beta.68
+
+## 0.1.0-beta.67
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-from-http-request@0.1.0-beta.67
+- @jpmorganchase/mosaic-schemas@0.1.0-beta.67
+- @jpmorganchase/mosaic-types@0.1.0-beta.67
+
+## 0.1.0-beta.66
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-from-http-request@0.1.0-beta.66
+- @jpmorganchase/mosaic-schemas@0.1.0-beta.66
+- @jpmorganchase/mosaic-types@0.1.0-beta.66
+
+## 0.1.0-beta.65
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-from-http-request@0.1.0-beta.65
+- @jpmorganchase/mosaic-schemas@0.1.0-beta.65
+- @jpmorganchase/mosaic-types@0.1.0-beta.65
+
 ## 0.1.0-beta.64
 
 ### Patch Changes