+++ packages/source-storybook/CHANGELOG.md
@@ -1,5 +1,64 @@
 # @jpmorganchase/mosaic-source-storybook
 
+## 0.1.0-beta.69
+
+### Patch Changes
+
+- f528d96b: Split Storybook endpoint into seperate paths
+
+  Previously we used a single url to access `stories.json` and the
+  linked story. Split this into
+
+  - `storiesUrl`: the full url to `stories.json`
+  - `storyUrlPrefix`: the url to prefix each story reference with
+
+  Refactored the `meta` of sources so we refer to the content preview as
+  'contentUrl`and the link we open in a seperate tab as`link`
+
+  - @jpmorganchase/mosaic-schemas@0.1.0-beta.69
+  - @jpmorganchase/mosaic-source-http@0.1.0-beta.69
+  - @jpmorganchase/mosaic-types@0.1.0-beta.69
+
+## 0.1.0-beta.68
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-schemas@0.1.0-beta.68
+- @jpmorganchase/mosaic-source-http@0.1.0-beta.68
+- @jpmorganchase/mosaic-types@0.1.0-beta.68
+
+## 0.1.0-beta.67
+
+### Patch Changes
+
+- 30e2f038: Align source APIs for Figma, Storybook and Readme
+
+  Updates to related sources so that have consistent APIs.
+
+  - `tags` is now visible in the `meta`, previously deleted
+  - use `meta.tags` rather than rely on a CSV copy in `meta.data.tags`
+  - move un-neccessary `meta.data` to config (e.g `source`)
+  - removed `additionalTags` and `additionalData` from Storybook source and use `meta` instead
+  - @jpmorganchase/mosaic-schemas@0.1.0-beta.67
+  - @jpmorganchase/mosaic-source-http@0.1.0-beta.67
+  - @jpmorganchase/mosaic-types@0.1.0-beta.67
+
+## 0.1.0-beta.66
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-schemas@0.1.0-beta.66
+- @jpmorganchase/mosaic-source-http@0.1.0-beta.66
+- @jpmorganchase/mosaic-types@0.1.0-beta.66
+
+## 0.1.0-beta.65
+
+### Patch Changes
+
+- @jpmorganchase/mosaic-schemas@0.1.0-beta.65
+- @jpmorganchase/mosaic-source-http@0.1.0-beta.65
+- @jpmorganchase/mosaic-types@0.1.0-beta.65
+
 ## 0.1.0-beta.64
 
 ### Patch Changes