+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @jpmorganchase/mosaic-theme
 
+## 0.1.0-beta.69
+
+## 0.1.0-beta.68
+
+## 0.1.0-beta.67
+
+## 0.1.0-beta.66
+
+## 0.1.0-beta.65
+
 ## 0.1.0-beta.64
 
 ## 0.1.0-beta.63